teacher_project/requirements.txt
streamlit 
python-dotenv 
langchain-google-genai 
google-generativeai 
reportlab
python-dateutil